+++ device/tree/src/property/interrupt_map.rs
@@ -32,23 +32,23 @@ impl SecondAnalyzed for InterruptMap {
             let mut words: Iter<'_, u32> = words.iter();
             let mut entries: Vec<Entry> = Vec::new();
             while words.len() != 0 {
-                let child_unit_address: usize = (0..address_cells)
+                let child_unit_address: u128 = (0..address_cells)
                     .map(|_| words.next().unwrap())
-                    .fold(0, |value, cell| (value << u32::BITS) + (*cell as usize));
-                let child_interrupt_specifier: usize = (0..interrupt_cells)
+                    .fold(0, |value, cell| (value << u32::BITS) + (*cell as u128));
+                let child_interrupt_specifier: u128 = (0..interrupt_cells)
                     .map(|_| words.next().unwrap())
-                    .fold(0, |value, cell| (value << u32::BITS) + (*cell as usize));
+                    .fold(0, |value, cell| (value << u32::BITS) + (*cell as u128));
                 let interrupt_parent: u32 = *words.next().unwrap();
                 let interrupt_parent_address_cells: usize =
                     second_analyzer.phandle_address_cells(interrupt_parent);
                 let interrupt_parent_interrupt_cells: usize =
                     second_analyzer.phandle_interrupt_cells(interrupt_parent);
-                let parent_unit_address: usize = (0..interrupt_parent_address_cells)
+                let parent_unit_address: u128 = (0..interrupt_parent_address_cells)
                     .map(|_| words.next().unwrap())
-                    .fold(0, |value, cell| (value << u32::BITS) + (*cell as usize));
-                let parent_interrupt_specifier: usize = (0..interrupt_parent_interrupt_cells)
+                    .fold(0, |value, cell| (value << u32::BITS) + (*cell as u128));
+                let parent_interrupt_specifier: u128 = (0..interrupt_parent_interrupt_cells)
                     .map(|_| words.next().unwrap())
-                    .fold(0, |value, cell| (value << u32::BITS) + (*cell as usize));
+                    .fold(0, |value, cell| (value << u32::BITS) + (*cell as u128));
                 entries.push(Entry {
                     child_unit_address,
                     child_interrupt_specifier,
@@ -66,9 +66,9 @@ impl SecondAnalyzed for InterruptMap {
 
 #[derive(Clone, Debug)]
 struct Entry {
-    child_unit_address: usize,
-    child_interrupt_specifier: usize,
+    child_unit_address: u128,
+    child_interrupt_specifier: u128,
     interrupt_parent: u32,
-    parent_unit_address: usize,
-    parent_interrupt_specifier: usize,
+    parent_unit_address: u128,
+    parent_interrupt_specifier: u128,
 }
